finish

+++ BE/biker-care/src/main/java/com/example/bikercare/order/dto/CartDto.java
@@ -9,5 +9,5 @@ public interface CartDto {
     Long getProductPrice();
     String getProductQuantity();
     String getProductImage();
-    Integer getProductId();
+    Long getProductId();
 }
